Break out IsoaltedStore into a separate file.  Man this was painful.
