+++ source/common/stats/isolated_store_impl.h
@@ -1,0 +1,81 @@
+#pragma once
+
+#include <string.h>
+
+#include <algorithm>
+#include <string>
+
+#include "common/common/utility.h"
+#include "common/stats/stats_impl.h"
+#include "common/stats/utility.h"
+
+namespace Envoy {
+namespace Stats {
+
+/**
+ * A stats cache template that is used by the isolated store.
+ */
+template <class Base> class IsolatedStatsCache {
+public:
+  typedef std::function<std::shared_ptr<Base>(const std::string& name)> Allocator;
+
+  IsolatedStatsCache(Allocator alloc) : alloc_(alloc) {}
+
+  Base& get(const std::string& name) {
+    auto stat = stats_.find(name);
+    if (stat != stats_.end()) {
+      return *stat->second;
+    }
+
+    std::shared_ptr<Base> new_stat = alloc_(name);
+    stats_.emplace(name, new_stat);
+    return *new_stat;
+  }
+
+  std::vector<std::shared_ptr<Base>> toVector() const {
+    std::vector<std::shared_ptr<Base>> vec;
+    vec.reserve(stats_.size());
+    for (auto& stat : stats_) {
+      vec.push_back(stat.second);
+    }
+
+    return vec;
+  }
+
+private:
+  std::unordered_map<std::string, std::shared_ptr<Base>> stats_;
+  Allocator alloc_;
+};
+
+class IsolatedStoreImpl : public Store {
+public:
+  IsolatedStoreImpl();
+
+  // Stats::Scope
+  Counter& counter(const std::string& name) override { return counters_.get(name); }
+  ScopePtr createScope(const std::string& name) override;
+  void deliverHistogramToSinks(const Histogram&, uint64_t) override {}
+  Gauge& gauge(const std::string& name) override { return gauges_.get(name); }
+  Histogram& histogram(const std::string& name) override {
+    Histogram& histogram = histograms_.get(name);
+    return histogram;
+  }
+  const Stats::StatsOptions& statsOptions() const override { return stats_options_; }
+
+  // Stats::Store
+  std::vector<CounterSharedPtr> counters() const override { return counters_.toVector(); }
+  std::vector<GaugeSharedPtr> gauges() const override { return gauges_.toVector(); }
+  std::vector<ParentHistogramSharedPtr> histograms() const override {
+    return std::vector<ParentHistogramSharedPtr>{};
+  }
+
+private:
+  HeapStatDataAllocator alloc_;
+  IsolatedStatsCache<Counter> counters_;
+  IsolatedStatsCache<Gauge> gauges_;
+  IsolatedStatsCache<Histogram> histograms_;
+  const StatsOptionsImpl stats_options_;
+};
+
+} // namespace Stats
+} // namespace Envoy
