+++ source/common/stats/stats_impl.h
@@ -33,65 +33,6 @@
 namespace Envoy {
 namespace Stats {
 
-/**
- * This structure is the backing memory for both CounterImpl and GaugeImpl. It is designed so that
- * it can be allocated from shared memory if needed.
- *
- * @note Due to name_ being variable size, sizeof(RawStatData) probably isn't useful. Use
- * RawStatData::structSize() or RawStatData::structSizeWithOptions() instead.
- */
-struct RawStatData {
-
-  /**
-   * Due to the flexible-array-length of name_, c-style allocation
-   * and initialization are necessary.
-   */
-  RawStatData() = delete;
-  ~RawStatData() = delete;
-
-  /**
-   * Returns the size of this struct, accounting for the length of name_
-   * and padding for alignment.
-   */
-  static uint64_t structSize(uint64_t name_size);
-
-  /**
-   * Wrapper for structSize, taking a StatsOptions struct.
-   * Required by BlockMemoryHashSet, which has the context to supply StatsOptions.
-   */
-  static uint64_t structSizeWithOptions(const StatsOptions& stats_options);
-
-  /**
-   * Initializes this object to have the specified key,
-   * a refcount of 1, and all other values zero. Required for the HeapRawStatDataAllocator, which
-   * does not expect stat name truncation. We pass in the number of bytes allocated in order to
-   * assert the copy is safe inline.
-   */
-  void initialize(absl::string_view key, const StatsOptions& stats_options);
-
-  /**
-   * Returns a hash of the key. This is required by BlockMemoryHashSet.
-   */
-  static uint64_t hash(absl::string_view key) { return HashUtil::xxHash64(key); }
-
-  /**
-   * Returns true if object is in use.
-   */
-  bool initialized() { return name_[0] != '\0'; }
-
-  /**
-   * Returns the name as a string_view with no truncation.
-   */
-  absl::string_view key() const { return absl::string_view(name_); }
-
-  std::atomic<uint64_t> value_;
-  std::atomic<uint64_t> pending_increment_;
-  std::atomic<uint16_t> flags_;
-  std::atomic<uint16_t> ref_count_;
-  std::atomic<uint32_t> unused_;
-  char name_[];
-};
-
 // Partially implements a StatDataAllocator, leaving alloc & free for subclasses.
 // We templatize on StatData rather than defining a virtual base StatData class
 // for performance reasons; stat increment is on the hot path.
@@ -280,71 +221,6 @@ private:
   // Although alloc() operations are called under existing locking, free() operations are made from
   // the destructors of the individual stat objects, which are not protected by locks.
   Thread::MutexBasicLockable mutex_;
-};
-
-/**
- * A stats cache template that is used by the isolated store.
- */
-template <class Base> class IsolatedStatsCache {
-public:
-  typedef std::function<std::shared_ptr<Base>(const std::string& name)> Allocator;
-
-  IsolatedStatsCache(Allocator alloc) : alloc_(alloc) {}
-
-  Base& get(const std::string& name) {
-    auto stat = stats_.find(name);
-    if (stat != stats_.end()) {
-      return *stat->second;
-    }
-
-    std::shared_ptr<Base> new_stat = alloc_(name);
-    stats_.emplace(name, new_stat);
-    return *new_stat;
-  }
-
-  std::vector<std::shared_ptr<Base>> toVector() const {
-    std::vector<std::shared_ptr<Base>> vec;
-    vec.reserve(stats_.size());
-    for (auto& stat : stats_) {
-      vec.push_back(stat.second);
-    }
-
-    return vec;
-  }
-
-private:
-  std::unordered_map<std::string, std::shared_ptr<Base>> stats_;
-  Allocator alloc_;
-};
-
-class IsolatedStoreImpl : public Store {
-public:
-  IsolatedStoreImpl();
-
-  // Stats::Scope
-  Counter& counter(const std::string& name) override { return counters_.get(name); }
-  ScopePtr createScope(const std::string& name) override;
-  void deliverHistogramToSinks(const Histogram&, uint64_t) override {}
-  Gauge& gauge(const std::string& name) override { return gauges_.get(name); }
-  Histogram& histogram(const std::string& name) override {
-    Histogram& histogram = histograms_.get(name);
-    return histogram;
-  }
-  const Stats::StatsOptions& statsOptions() const override { return stats_options_; }
-
-  // Stats::Store
-  std::vector<CounterSharedPtr> counters() const override { return counters_.toVector(); }
-  std::vector<GaugeSharedPtr> gauges() const override { return gauges_.toVector(); }
-  std::vector<ParentHistogramSharedPtr> histograms() const override {
-    return std::vector<ParentHistogramSharedPtr>{};
-  }
-
-private:
-  HeapStatDataAllocator alloc_;
-  IsolatedStatsCache<Counter> counters_;
-  IsolatedStatsCache<Gauge> gauges_;
-  IsolatedStatsCache<Histogram> histograms_;
-  const StatsOptionsImpl stats_options_;
 };
 
 } // namespace Stats
