+++ source/common/upstream/upstream_impl.h
@@ -33,6 +33,7 @@
 #include "common/config/metadata.h"
 #include "common/config/well_known_names.h"
 #include "common/network/utility.h"
+#include "common/stats/isolated_store_impl.h"
 #include "common/stats/stats_impl.h"
 #include "common/upstream/load_balancer_impl.h"
 #include "common/upstream/locality.h"
