+++ source/server/hot_restart_impl.cc
@@ -18,6 +18,7 @@
 #include "common/common/lock_guard.h"
 #include "common/common/utility.h"
 #include "common/network/utility.h"
+#include "common/stats/raw_stat_data.h"
 
 #include "absl/strings/string_view.h"
