+++ test/common/filesystem/BUILD
@@ -17,6 +17,7 @@ envoy_cc_test(
         "//source/common/event:dispatcher_includes",
         "//source/common/event:dispatcher_lib",
         "//source/common/filesystem:filesystem_lib",
+        "//source/common/stats:isolated_store_lib",
         "//source/common/stats:stats_lib",
         "//test/mocks/api:api_mocks",
         "//test/mocks/event:event_mocks",
