+++ test/common/filesystem/filesystem_impl_test.cc
@@ -6,6 +6,7 @@
 #include "common/common/thread.h"
 #include "common/event/dispatcher_impl.h"
 #include "common/filesystem/filesystem_impl.h"
+#include "common/stats/isolated_store_impl.h"
 #include "common/stats/stats_impl.h"
 
 #include "test/mocks/api/mocks.h"
