+++ test/common/grpc/google_async_client_impl_test.cc
@@ -1,5 +1,6 @@
 #include "common/event/dispatcher_impl.h"
 #include "common/grpc/google_async_client_impl.h"
+#include "common/stats/isolated_store_impl.h"
 #include "common/stats/stats_impl.h"
 
 #include "test/mocks/grpc/mocks.h"
