+++ test/common/runtime/BUILD
@@ -21,6 +21,7 @@ envoy_cc_test(
     data = glob(["test_data/**"]) + ["filesystem_setup.sh"],
     deps = [
         "//source/common/runtime:runtime_lib",
+        "//source/common/stats:isolated_store_lib",
         "//source/common/stats:stats_lib",
         "//test/mocks/api:api_mocks",
         "//test/mocks/event:event_mocks",
