+++ test/common/runtime/runtime_impl_test.cc
@@ -2,6 +2,7 @@
 #include <string>
 
 #include "common/runtime/runtime_impl.h"
+#include "common/stats/isolated_store_impl.h"
 #include "common/stats/stats_impl.h"
 
 #include "test/mocks/api/mocks.h"
