+++ test/mocks/network/BUILD
@@ -22,6 +22,7 @@ envoy_cc_mock(
         "//include/envoy/server:listener_manager_interface",
         "//source/common/network:address_lib",
         "//source/common/network:utility_lib",
+        "//source/common/stats:isolated_store_lib",
         "//test/mocks/event:event_mocks",
         "//test/test_common:printers_lib",
         "@envoy_api//envoy/api/v2/core:base_cc",
